python-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pdfplumber==0.9.0
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.4.2